@@ -20,7 +20,7 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   flutter_native_splash: ^2.4.4
-  sentry_flutter: ^8.12.0
+  sentry_flutter: ^9.0.0
   smooth_page_indicator: ^1.2.0+3
 
   # The following adds the Cupertino Icons font to your application.
@@ -53,16 +53,18 @@ dependencies:
   email_validator: ^3.0.0
   google_sign_in: ^6.2.2
   sign_in_with_apple: ^7.0.1
-  flutter_platform_widgets: ^8.0.0
+  flutter_platform_widgets: ^9.0.0
   dynamic_color: ^1.7.0
   connectivity_plus: ^6.1.1
   toastification: ^3.0.2
-  health: ^12.1.0
+  health: ^13.0.1
   device_info_plus: ^11.3.3
   location: ^8.0.0
   logger: ^2.5.0
   async: ^2.12.0
   easy_localization: ^3.0.7+1
+  flutter_typeahead: ^5.2.0
+  video_player: ^2.10.0
   # awesome_notifications: ^0.10.0
   # google_mobile_ads: ^5.2.0
 
@@ -76,11 +78,11 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^5.0.0
+  flutter_lints: ^6.0.0
+  lints: ^6.0.0
+
   json_serializable: ^6.9.4
   build_runner: ^2.4.15
-  lints: ^5.1.1
-
   # For information on the generic Dart part of this file, see the
   # following page: https://dart.dev/tools/pub/pubspec
 
@@ -276,3 +278,7 @@ flutter_native_splash:
   #info_plist_files:
   #  - 'ios/Runner/Info-Debug.plist'
   #  - 'ios/Runner/Info-Release.plist'
+
+scripts:
+  gen: dart run build_runner build --delete-conflicting-outputs
+  svg: dart run vector_graphics_compiler --input-dir ./assets/icons --out-dir ./assets/icons
